made amendements to other location description page
