+++ server/views/blockage/location-description-optional.html
